Don't read locale data from DATA - instead load DATA into a string at module load time
Greg Oschwald reported a bug where attempting to eval the locale data for a
locale would randomly throw an error. Looking at the error, it seems like we
somehow read a bad chunk of data from DATA, overlapping two
locales. Presumably _something_ is messing with the DATA handle, but figuring
out what that is in a complicated program is probably impossible, so we'll try
this alternative approach.