Accepted a tainted code passed to DateTime::Locale->load and untaint it
I don't think there were any security issues without this, as the load method
would only attempt to load data from disk if the code was already known. This
means that attempting to pass in any sort of arbitrary path to the load method
would simply result in an error that the code was invalid, and there would be
no attempt to read from disk.